Refactored creation dialog to ResponsiveAppBar
